Move all state into state.ts, extract util.ts
