--- app/javascript/projects/project.ts
@@ -1,8 +1,0 @@
-import { Layer } from './layers'
-
-export interface Project {
-  name: string
-  layers: Record<number, Layer>
-  allLayers: number[]
-  selectedLayer?: number
-}
