+++ app/javascript/projects/project_editor.tsx
@@ -3,7 +3,7 @@ import './project_editor.css'
 import { LayerPalette } from './layer_palette'
 import { Toolbar } from './toolbar'
 import { reifyLayer } from './reify_layer'
-import { Project } from './project'
+import { Project } from './state'
 import { Sidebar, CollapsedSidebar } from './sidebar'
 import { MapView } from './map_view'
 import { DBModels } from './db_models'
